# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POC OCR + LLM — Notas Fiscais",
    "description": "POC do TCC — extração automática de dados em notas fiscais com OCR (Tesseract) e LLM. Fluxo: documento → OCR → LLM → validação sistêmica → validação humana → MongoDB. Campos: issuer_name, cnpj, issue_date, invoice_number, total_value.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Disponibilidade da API e do MongoDB",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Enviar documento e processar (síncrono)",
        "operationId": "create_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Consultar documento",
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Correção / confirmação humana",
        "operationId": "update_document_api_v1_documents__doc_id__put",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}/reprocess": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Reprocessar a partir do arquivo original",
        "operationId": "reprocess_document_api_v1_documents__doc_id__reprocess_post",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Nota fiscal em PDF, PNG ou JPG/JPEG"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_document_api_v1_documents_post"
      },
      "Correction": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "old_value": {
            "title": "Old Value"
          },
          "new_value": {
            "title": "New Value"
          },
          "corrected_at": {
            "type": "string",
            "title": "Corrected At"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "human"
          }
        },
        "type": "object",
        "required": [
          "field",
          "corrected_at"
        ],
        "title": "Correction"
      },
      "DocumentOut": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "source_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pages"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "extracted_data": {
            "$ref": "#/components/schemas/ExtractedFields"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationInfo"
          },
          "llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingInfo"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "file_name",
          "file_path",
          "mime_type",
          "file_size_bytes",
          "status",
          "extracted_data",
          "validation",
          "processing",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentOut"
      },
      "ErrorInfo": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "at": {
            "type": "string",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "message",
          "at"
        ],
        "title": "ErrorInfo"
      },
      "ExtractedFields": {
        "properties": {
          "issuer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "total_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExtractedFields",
        "description": "Dados estruturados dos 5 campos essenciais (armazenamento / saída / edição).\n\nTodos os campos são opcionais: ausência de evidência ⇒ `null` (o sistema\nnunca preenche valores presumidos — ESPECIFICACAO.md §8)."
      },
      "ExtractedFieldsUpdate": {
        "properties": {
          "issuer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "total_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExtractedFieldsUpdate",
        "description": "Correção humana parcial: só os campos presentes no corpo são alterados."
      },
      "FieldCheck": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "FieldCheck"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMInfo": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "raw_response_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Response Excerpt"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "attempts"
        ],
        "title": "LLMInfo"
      },
      "ProcessingInfo": {
        "properties": {
          "ocr_duration_ms": {
            "type": "integer",
            "title": "Ocr Duration Ms",
            "default": 0
          },
          "llm_duration_ms": {
            "type": "integer",
            "title": "Llm Duration Ms",
            "default": 0
          },
          "total_duration_ms": {
            "type": "integer",
            "title": "Total Duration Ms",
            "default": 0
          },
          "reprocess_count": {
            "type": "integer",
            "title": "Reprocess Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "ProcessingInfo"
      },
      "UpdateRequest": {
        "properties": {
          "extracted_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractedFieldsUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "validated": {
            "type": "boolean",
            "title": "Validated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationInfo": {
        "properties": {
          "schema_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Valid"
          },
          "field_checks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldCheck"
            },
            "type": "object",
            "title": "Field Checks"
          },
          "validated": {
            "type": "boolean",
            "title": "Validated",
            "default": false
          },
          "validated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated By"
          },
          "validated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "corrections": {
            "items": {
              "$ref": "#/components/schemas/Correction"
            },
            "type": "array",
            "title": "Corrections"
          }
        },
        "type": "object",
        "title": "ValidationInfo"
      }
    }
  }
}